fix: cleaned up code, removed unescessary redux implementation, fixed small plan reload bug
